MNT: Move mpl keybinding fix to 'not building'
conda-build-all is not correctly determining the build string so the
circle-ci builds are taking forever.  We have the recipe built already,
so if there are new commits we can just move the recipe back in to pyall
and it will build.  Then we can move it back in to not-building so it
does not slow down the circle-ci builds