MNT: Move stuff back out of x.x
I moved stuff into x.x because those recipes have the "numpy x.x" build/runtime
requirement that was not being handled properly.  I now know that this is because
I was trying to build against numpy 1.11 which is not available behind the nsls2
firewall.  Adding numpy 1.10 will probably fix this...