Move lmfit pin from scikit-beam to pyxrf.
This leaves it up to the individual beamlines whether to pin lmfit to 0.8.3 and
get working xrf code in scikit-beam or accept broken xrf code in scikit-beam in
order to use Larch and modern lmfit.

The ideal fix would be to update the xrf code in scikit-beam, but that update
is not straightforward, and we cannot wait for it.